+++ flycheck.el
@@ -11866,11 +11866,17 @@ This syntax checker needs Rust 1.18 or newer.  See URL
   :modes (rust-mode rust-ts-mode)
   :predicate flycheck-buffer-saved-p)
 
+(flycheck-def-args-var flycheck-rust-clippy-args rust-clippy
+  :package-version '(flycheck . "36.1"))
+
 (flycheck-define-checker rust-clippy
   "A Rust syntax checker using clippy.
 
 See URL `https://github.com/rust-lang-nursery/rust-clippy'."
-  :command ("cargo" "clippy" "--message-format=json")
+  :command ("cargo"
+            "clippy"
+            "--message-format=json"
+            (eval flycheck-rust-clippy-args))
   :error-parser flycheck-parse-cargo-rustc
   :error-filter flycheck-rust-error-filter
   :error-explainer flycheck-rust-error-explainer